Improved installation procedure

+++ setup.py
@@ -25,25 +25,28 @@ import setuptools
 from setuptools import setup, Extension, find_packages
 from setuptools.command.build_ext import build_ext
 from distutils.version import LooseVersion
-from shutil import rmtree
+
 
 # create file structure for testing and MTEX support
 # not required otherwise
 MAIN_DIR = os.getcwd()  # directory in which repository is cloned
 WORK_DIR = os.path.expanduser('~') + '/.kanapy'  # working directory for temporary files
-if os.path.exists(WORK_DIR):
-    print(f'Removing existing working directory {WORK_DIR}')
-    rmtree(WORK_DIR)
-
-os.makedirs(WORK_DIR)
-os.makedirs(WORK_DIR + '/tests')
-os.system(f'cp {MAIN_DIR}/tests/unitTest_ODF_MDF_reconstruction.m {WORK_DIR}/tests')
-pathDict = {'MAIN_DIR': MAIN_DIR, 'MTEXpath': MAIN_DIR+'/libs/mtex', 'MATLABpath': None}
 path_path = WORK_DIR + '/PATHS.json'
-
-
-with open(path_path, 'w') as outfile:
-    json.dump(pathDict, outfile, indent=2)
+path_dict = {'MAIN_DIR': MAIN_DIR,
+             'MTEXpath': MAIN_DIR+'/libs/mtex',
+             'MATLABpath': None}
+if os.path.exists(WORK_DIR):
+    if not os.path.exists(WORK_DIR + '/tests'):
+        os.makedirs(WORK_DIR + '/tests')
+    if not os.path.exists(path_path):
+        with open(path_path, 'w') as outfile:
+            json.dump(path_dict, outfile, indent=2)
+else:
+    os.makedirs(WORK_DIR)
+    os.makedirs(WORK_DIR + '/tests')
+    os.system(f'cp {MAIN_DIR}/tests/unitTest_ODF_MDF_reconstruction.m {WORK_DIR}/tests')
+    with open(path_path, 'w') as outfile:
+        json.dump(path_dict, outfile, indent=2)
     
 """The setup script."""
